requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
apispec==6.6.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.8.0
comm==0.2.2
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
gast==0.5.4
google-auth==2.29.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyzmq==25.1.2
requests==2.31.0
rich==13.7.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
webargs==8.4.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
